# Fix deadlock when using Iter/IterBuffered