--- a/concurrent_map_test.go
+++ b/concurrent_map_test.go
@@ -467,3 +467,108 @@ func TestKeysWhenRemoving(t *testing.T) {
 		}
 	}
 }
+
+//
+func TestUnDrainedIter(t *testing.T) {
+	m := New()
+	// Insert 100 elements.
+	Total := 100
+	for i := 0; i < Total; i++ {
+		m.Set(strconv.Itoa(i), Animal{strconv.Itoa(i)})
+	}
+	counter := 0
+	// Iterate over elements.
+	ch := m.Iter()
+	for item := range ch {
+		val := item.Val
+
+		if val == nil {
+			t.Error("Expecting an object.")
+		}
+		counter++
+		if counter == 42 {
+			break
+		}
+	}
+	for i := Total; i < 2*Total; i++ {
+		m.Set(strconv.Itoa(i), Animal{strconv.Itoa(i)})
+	}
+	for item := range ch {
+		val := item.Val
+
+		if val == nil {
+			t.Error("Expecting an object.")
+		}
+		counter++
+	}
+
+	if counter != 100 {
+		t.Error("We should have been right where we stopped")
+	}
+
+	counter = 0
+	for item := range m.IterBuffered() {
+		val := item.Val
+
+		if val == nil {
+			t.Error("Expecting an object.")
+		}
+		counter++
+	}
+
+	if counter != 200 {
+		t.Error("We should have counted 200 elements.")
+	}
+}
+
+func TestUnDrainedIterBuffered(t *testing.T) {
+	m := New()
+	// Insert 100 elements.
+	Total := 100
+	for i := 0; i < Total; i++ {
+		m.Set(strconv.Itoa(i), Animal{strconv.Itoa(i)})
+	}
+	counter := 0
+	// Iterate over elements.
+	ch := m.IterBuffered()
+	for item := range ch {
+		val := item.Val
+
+		if val == nil {
+			t.Error("Expecting an object.")
+		}
+		counter++
+		if counter == 42 {
+			break
+		}
+	}
+	for i := Total; i < 2*Total; i++ {
+		m.Set(strconv.Itoa(i), Animal{strconv.Itoa(i)})
+	}
+	for item := range ch {
+		val := item.Val
+
+		if val == nil {
+			t.Error("Expecting an object.")
+		}
+		counter++
+	}
+
+	if counter != 100 {
+		t.Error("We should have been right where we stopped")
+	}
+
+	counter = 0
+	for item := range m.IterBuffered() {
+		val := item.Val
+
+		if val == nil {
+			t.Error("Expecting an object.")
+		}
+		counter++
+	}
+
+	if counter != 200 {
+		t.Error("We should have counted 200 elements.")
+	}
+}