@flyingfoxlee 2017-03-08
It also fix issue #42 and passes the test proposed by @teou.
It also fixes #46, @Glacier759  now you can delete element when using `Iter` function happily~