@orcaman 2017-03-09
@flyingfoxlee nice work! I would like to merge this PR, but I'm a bit  concerned about the performance. When you run `go test -bench=.`, do you see a significant difference between the master and this version?