@flyingfoxlee 2017-03-10
On my Macbook Air, `go test -bench=.` returns as follows
Before:
BenchmarkItems-4                                     200           6314318 ns/op
BenchmarkMarshalJson-4                               100          18237536 ns/op
BenchmarkStrconv-4                              20000000                64.1 ns/op
BenchmarkSingleInsertAbsent-4                    1000000              1171 ns/op
BenchmarkSingleInsertPresent-4                  10000000               151 ns/op
BenchmarkMultiInsertDifferent_1_Shard-4          1000000              3377 ns/op
BenchmarkMultiInsertDifferent_16_Shard-4         1000000              3334 ns/op
BenchmarkMultiInsertDifferent_32_Shard-4         1000000              3084 ns/op
BenchmarkMultiInsertDifferent_256_Shard-4         500000              4113 ns/op
BenchmarkMultiInsertSame-4                       1000000              1685 ns/op
BenchmarkMultiGetSame-4                          2000000               663 ns/op
BenchmarkMultiGetSetDifferent_1_Shard-4           500000              4202 ns/op
BenchmarkMultiGetSetDifferent_16_Shard-4          500000              4207 ns/op
BenchmarkMultiGetSetDifferent_32_Shard-4          500000              4272 ns/op
BenchmarkMultiGetSetDifferent_256_Shard-4         500000              4077 ns/op
BenchmarkMultiGetSetBlock_1_Shard-4               500000              2732 ns/op
BenchmarkMultiGetSetBlock_16_Shard-4              500000              2726 ns/op
BenchmarkMultiGetSetBlock_32_Shard-4              500000              2563 ns/op
BenchmarkMultiGetSetBlock_256_Shard-4            1000000              2350 ns/op
BenchmarkKeys-4                                      500           3115590 ns/op

After:
BenchmarkItems-4                                     200           7822025 ns/op
BenchmarkMarshalJson-4                               100          18966387 ns/op
BenchmarkStrconv-4                              20000000                69.2 ns/op
BenchmarkSingleInsertAbsent-4                    1000000              1197 ns/op
BenchmarkSingleInsertPresent-4                  10000000               151 ns/op
BenchmarkMultiInsertDifferent_1_Shard-4           500000              3093 ns/op
BenchmarkMultiInsertDifferent_16_Shard-4         1000000              3190 ns/op
BenchmarkMultiInsertDifferent_32_Shard-4          500000              3066 ns/op
BenchmarkMultiInsertDifferent_256_Shard-4         300000              3668 ns/op
BenchmarkMultiInsertSame-4                       1000000              1647 ns/op
BenchmarkMultiGetSame-4                          2000000               620 ns/op
BenchmarkMultiGetSetDifferent_1_Shard-4           300000              3738 ns/op
BenchmarkMultiGetSetDifferent_16_Shard-4          500000              4338 ns/op
BenchmarkMultiGetSetDifferent_32_Shard-4          500000              4181 ns/op
BenchmarkMultiGetSetDifferent_256_Shard-4         300000              3754 ns/op
BenchmarkMultiGetSetBlock_1_Shard-4               500000              2866 ns/op
BenchmarkMultiGetSetBlock_16_Shard-4              500000              2758 ns/op
BenchmarkMultiGetSetBlock_32_Shard-4              500000              2520 ns/op
BenchmarkMultiGetSetBlock_256_Shard-4            1000000              2422 ns/op
BenchmarkKeys-4                                      500           3115989 ns/op